+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Changelog
+
+## 0.0.3 - 2024-05-04
+
+- Adding Dart `args` plugin.
+- Check new version available.
+- Adding recursive option.
+
 ## 0.0.2 - 2024-05-04
 
 - Adding Kotlin support.